+++ src/Aks/Aks/ChangeLog.md
@@ -25,6 +25,7 @@
 * Added parameter `-NodeMaxSurge` for `New-AzAksCluster`, `-MaxSurge` for `New-AzAksNodePool` and `Update-AzAksNodePool`
 * Added parameter `-PPG` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-SpotMaxPrice` for `New-AzAksNodePool`
+* Added parameter `-EnableFIPS` for `New-AzAksCluster` and `New-AzAksNodePool`
 
 ## Version 5.1.0
 * Bumped API version to 2022-09-01